Changing DB
The DBName created on Google Cloud SQL DOES recognise Letter Case, so the name is Changed to work as exepcted.
Doing some test in Google Cloud: npm install (not needed 'cause all libraries are committed, a worng practice sometimes) + npm start, to execute in a controlled environment, on the cloud.
This time (on the cloud) DB file will be using the Database from the own Cloud (SQL) including an External Static IP on the InstanceVM and grant access to the Database from it (external IP).